+++ app/backup.sh
@@ -29,13 +29,7 @@ build_restic_cacert_args
 # Clear log files
 rm -f "${LAST_LOGFILE}" "${LAST_MAIL_LOGFILE}"
 
-# Check if pre-backup script exists and execute it
-if [ -f "/hooks/pre-backup.sh" ]; then
-	log "🚀 Starting pre-backup script..."
-	/hooks/pre-backup.sh
-else
-	log "ℹ️ Pre-backup script not found..."
-fi
+run_hook "pre-backup"
 
 # Record start time
 start=$(date +%s)
@@ -138,12 +132,6 @@ if [ -n "${MAILX_RCPT}" ] && {
 	fi
 fi
 
-# Check if post-backup script exists and execute it
-if [ -f "/hooks/post-backup.sh" ]; then
-	log "🚀 Starting post-backup script..."
-	/hooks/post-backup.sh "$backupRC"
-else
-	log "ℹ️ Post-backup script not found..."
-fi
+run_hook "post-backup" "$backupRC" || true
 
 exit "$backupRC"